% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osqp.R
\name{OSQP_Model}
\alias{OSQP_Model}
\title{OSQP Model Class (S7)}
\usage{
OSQP_Model(work)
}
\description{
An S7 class representing an OSQP solver model. Methods are accessed via
computed properties using the \code{@} operator.
}
\keyword{internal}
